# packages/dashboard-ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memex Core API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/resources/{path}": {
      "get": {
        "summary": "Get Resource",
        "description": "Retrieve a raw resource (file) from the filestore.",
        "operationId": "get_resource_api_v1_resources__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{id}/lineage": {
      "get": {
        "summary": "Get Note Lineage",
        "description": "Get the lineage of a note.",
        "operationId": "get_note_lineage_api_v1_notes__id__lineage_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LineageDirection",
              "default": "upstream"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "Depth"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineageDirection": {
        "type": "string",
        "enum": [
          "upstream",
          "downstream",
          "both"
        ],
        "title": "LineageDirection",
        "description": "Direction of lineage traversal."
      },
      "LineageResponse": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "description": "The type of the entity (mental_model, observation, etc.)"
          },
          "entity": {
            "additionalProperties": true,
            "type": "object",
            "title": "Entity",
            "description": "The full DTO of the entity."
          },
          "derived_from": {
            "items": {
              "$ref": "#/components/schemas/LineageResponse"
            },
            "type": "array",
            "title": "Derived From",
            "description": "List of entities this entity is derived from (upstream) or that are derived from it (downstream)."
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity"
        ],
        "title": "LineageResponse",
        "description": "Recursive schema for lineage response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}